refactor(settings): アニメミュートのフォーム処理を共有ヘルパーへ集約
updateAnimeMute/removeAnimeMute のformDataパースと検証を
actions.ts の updateAnimeMuteAction/removeAnimeMuteAction へ移し、
ルートはセッション確認とヘルパー転送のみの薄い実装にする。
parsePositiveInt も actions.ts へ移設して定義元を一本化。
